+++ packages/sdks/nextjs-sdk/src/shared/index.ts
@@ -1,9 +1,3 @@
-// workaround for TS issue https://github.com/microsoft/TypeScript/issues/42873
-// eslint-disable-next-line
-import type * as _1 from '@descope/web-component';
-// eslint-disable-next-line
-import type * as _2 from '@descope/core-js-sdk';
-
 export { default as AuthProvider } from './AuthProvider';
 export * from './DescopeFlows';
 export * from './DescopeWidgets';